+++ cmd_line/source/cl_tok.c
@@ -1,7 +1,5 @@
 #include <stdlib.h>
 #include <string.h>
-#include "cl_common.h"
-#include "cl_tok.h"
 #include "cl_tok_p.h"
 
 #ifndef CL_TOK_LENGTH_MAX
@@ -23,8 +21,8 @@ static cl_bool is_char_digit(char c) {
     return CL_FALSE;
 }
 
-size_t cl_tok_get_length(cl_tok *p) {
-    size_t length;
+int cl_tok_get_length(cl_tok *p) {
+    int length;
     if (NULL == p) return 0;
 
     for (length = 0; *p != cl_tok_separator; length++, p++);
@@ -38,7 +36,7 @@ const char *cl_tok_get_value(cl_tok *p) {
 }
 
 cl_bool cl_tok_equals(cl_tok *p, const char *value) {
-    size_t i, len;
+    int i, len;
 
     /* check if these pointers are the same */
     if (value == p) return CL_TRUE;
@@ -63,7 +61,7 @@ cl_bool cl_tok_equals(cl_tok *p, const char *value) {
 }
 
 cl_bool cl_tok_is_numeric(cl_tok *p) {
-    size_t i, len;
+    int i, len;
     cl_bool dec_found;
 
     /* check if p is NULL */
@@ -113,7 +111,7 @@ cl_bool cl_tok_is_numeric(cl_tok *p) {
 }
 
 cl_bool cl_tok_is_switch(cl_tok* p) {
-    size_t len = 0;
+    int len = 0;
 
     /* check for a null pointer */
     if (NULL == p) return CL_FALSE;
